reelsense_backend/requirements.txt
# Core dependencies
pandas==2.0.3
numpy<2.0.0
scikit-learn>=1.3.0
scipy==1.11.1

# Matrix factorization
scikit-surprise>=1.1.3

# Visualization (for training phase)
matplotlib==3.7.2
seaborn==0.12.2

# Progress bars
tqdm==4.65.0

# Model serialization
joblib==1.3.1

# API (for next phase)
fastapi==0.100.0
uvicorn[standard]==0.23.0
pydantic==2.0.3

# Database (for next phase)
psycopg2-binary==2.9.6
redis==4.6.0

# Authentication (for next phase)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.0
pytest-cov==4.1.0
